requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
branca==0.7.0
certifi==2023.11.17
cfgv==3.3.1
charset-normalizer==3.3.2
distlib==0.3.6
Django==4.2.1
django-debug-toolbar==4.1.0
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
filelock==3.12.2
folium==0.15.1
identify==2.5.24
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.3
nodeenv==1.8.0
numpy==1.24.4
platformdirs==3.8.0
pre-commit==3.3.3
pyflame==0.3.1
PyJWT==1.7.1
pytz==2023.3
PyYAML==6.0
redis==4.5.5
requests==2.31.0
sqlparse==0.4.4
typing-extensions==4.6.2
urllib3==2.1.0
virtualenv==20.23.1
xyzservices==2023.10.1
